Move norm implementation into nda/linalg/norm.hpp
