+++ c++/nda/basic_functions.hpp
@@ -18,7 +18,6 @@
 
 #include "basic_array.hpp"
 #include "traits.hpp"
-#include "blas.hpp"
 
 namespace nda {
 
@@ -65,52 +64,6 @@ namespace nda {
   template <typename T, std::integral... Ints>
   auto ones(Ints... i) {
     return ones<T>(std::array<long, sizeof...(Ints)>{i...});
-  }
-
-  // --------------------------- norm ------------------------
-
-  /**
-   * Calculate the p-norm of an array x with scalar values and rank one:
-   *
-   *   norm(x) = sum(abs(x)^ord)^(1./ord)
-   *
-   * with the special cases (following numpy.linalg.norm convention)
-   *
-   *   norm(x, 0.0)  = number of non-zero elements
-   *   norm(x, inf)  = max_element(abs(x))
-   *   norm(x, -inf) = min_element(abs(x))
-   * 
-   * @param x The array to calculate the norm of
-   * @param p The order of the norm [default=2.0]
-   * @tparam A The type of the array
-   * @return The norm as a double
-   */
-  template <ArrayOfRank<1> A>
-  double norm(A const &x, double p = 2.0)
-    requires(Scalar<get_value_t<A>>)
-  {
-
-    if (p == 2.0) [[likely]] {
-      if constexpr (MemoryArray<A>)
-        return std::sqrt(std::real(nda::blas::dotc(x, x)));
-      else
-        return norm(make_regular(x));
-    } else if (p == 1.0) {
-      return sum(abs(x));
-    } else if (p == 0.0) {
-      // return std::count_if(a.begin(), a.end(), [](S s) { return s != S{0}; }); Fails for nda::expr
-      long count = 0;
-      for (long i = 0; i < x.size(); ++i) {
-        if (x(i) != get_value_t<A>{0}) ++count;
-      }
-      return double(count);
-    } else if (p == std::numeric_limits<double>::infinity()) {
-      return max_element(abs(x));
-    } else if (p == -std::numeric_limits<double>::infinity()) {
-      return min_element(abs(x));
-    } else {
-      return std::pow(sum(pow(abs(x), p)), 1.0 / p);
-    }
   }
 
   // --------------------------- arange ------------------------
